refactor: 인박스 목록 아이템 컴포넌트 분리 (#12)
